Mage: add SetBuildVersion target and remove use of VERSION environment variable
This removes the use of the VERSION environment variable to be used
when building a release of the bot. The SetBuildVersion target can
be used to set a version (eg 3.5.1) which will then be used to build
a binary or Docker image.

If no version is specified, the first 6 characters of the latest Git
commit hash are used instead.

Signed-off-by: Tom <tom@tdpain.net>