Bio images can now be removed with the bio site tool
Empty strings signify a deleted value. Bio image URLs are passed
through `url.ParseRequestURI` to ensure they're in a correct value,
but this will thrown an error if an empty string is passed to it.

This fixes that by adding a switch that means input strings are
only parsed if they're *not* empty strings.

Signed-off-by: AKP <tom@tdpain.net>